pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "call-evaluation-project"
version = "0.1.0"
description = "Transcript analysis for profanity, compliance, and call metrics."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.52.0",
  "pandas>=2.2.3",
  "plotly>=5.24.1",
  "pydantic>=2.9.2",
  "PyYAML>=6.0.2",
  "python-dotenv>=1.0.1",
  "streamlit>=1.39.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
